requirements/base.txt
bcrypt>=3.1.4
colorlog>=3.1.0
coverage>=4.5
Django==2.0.2
django-anymail==1.3
django-htmlmin>=0.10.0
django-webtest>=1.9.2
factory-boy>=2.10.0
psycopg2>=2.7.3.2
PyYAML>=3.12
raven>=6.5.0
requests>=2.18.4
